Module ncomm_publishers_and_subscribers::local

Local Publishers and Subscribers

Local Publishers and Subscribers utilize some combination of primitives from the standard library to enable the sharing of data between publishers and subscribers

Structs

LocalBufferedSubscriber

Local Subscriber that stores incoming data into a buffer for processing all at once

LocalMappedSubscriber

Local subscriber that maps incoming data to into a location in a hashmap allowing the subscriber to maintain a number of pieces of data at once.

LocalMappedTTLSubscriber

Local subscriber that maps incoming data to into a location in a hashmap while specifying a time-to-live for pieces of data contained in the hashmap

LocalPublisher

Local Publisher that utilizes a crossbeam multi publisher multi subscriber to send data

LocalSubscriber

Local Subscriber that utilizes a crossbeam multi subscriber channel to receive data from a local publisher

LocalTTLSubscriber

Local subscriber where data has a specific time-to-live and will decay after the lifetime has passed