Module ncomm_publishers_and_subscribers::tcp

A Network Tcp-Based Publisher and Subscriber

The TCP Publisher sends data via a TCP Stream to a bound listener on the Subscriber end

Structs

TcpBufferedSubscriber

A Tcp Subscriber that stores incoming data into a clearable buffer

TcpMappedSubscriber

A Tcp Subscriber that maps incoming data to its IP address.

TcpMappedTTLSubscriber

A Tcp Subscriber that stores incoming data stored by IP Address with a given time-to-live for stored data.

TcpPublisher

A Tcp Publisher that publishes data via packing the data according to the data’s Packable implementation

TcpSubscriber

A Tcp Subscriber that is set to nonblocking and and listens to incoming data. If data comes from an unknown IP address, the subscriber will reject the incoming data.

TcpTTLSubscriber

A Tcp Subscriber that subscribes to a TCP stream keeping data for specified time-to-live

Enums

TcpPublishError

An Error when attempting to publish data over a Tcp Publisher